% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange}
\alias{arrange}
\title{Arrange rows by variables}
\usage{
arrange(.data, ...)
}
\arguments{
\item{.data}{A \code{data.frame}.}

\item{...}{A comma separated vector of unquoted name(s) to order the data by.}
}
\value{
A \code{data.frame}.
}
\description{
Order rows of a \code{data.frame} by an expression involving its variables.
}
\examples{
arrange(mtcars, mpg)
mtcars \%>\% arrange(mpg)
mtcars \%>\% arrange(cyl, mpg)

}
