% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between.R
\name{between}
\alias{between}
\title{Do values in a numeric vector fall in specified range?}
\usage{
between(x, left, right)
}
\arguments{
\item{x}{A \code{numeric} vector of values.}

\item{left, right}{Boundary values.}
}
\value{
A \code{logical} vector the same length as \code{x}.
}
\description{
This is a shortcut for ‘x >= left & x <= right’.
}
\examples{
between(1:12, 7, 9)

x <- rnorm(1e2)
x[between(x, -1, 1)]

}
