% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{Filter}
\usage{
filter(.data, ...)
}
\arguments{
\item{.data}{A `data.frame`.}

\item{...}{Expressions used to filter the data by.}
}
\description{
Filter
}
\examples{
filter(mtcars, am == 1)
mtcars \%>\% filter(cyl == 4)
mtcars \%>\% filter(cyl <= 5 & am > 0)
mtcars \%>\% filter(cyl == 4 | cyl == 8)
mtcars \%>\% filter(!(cyl \%in\% c(4, 6)), am != 0)

}
