% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins_filtering.R
\name{filter_joins}
\alias{filter_joins}
\alias{anti_join}
\title{Filtering joins filter rows from \code{x} based on the presence or absence of
matches in \code{y}:}
\usage{
anti_join(x, y, by = NULL)
}
\description{
\itemize{
\item \code{semi_join()} return all rows from \code{x} with a match in \code{y}.
\item \code{anti_join()} return all rows from \code{x} with\emph{out} a match in \code{y}.
}
}
