% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_cols.R
\name{group_cols}
\alias{group_cols}
\title{Select Grouping Variables}
\usage{
group_cols()
}
\description{
This selection helper matches grouping variables. It can be used within \code{\link[=select]{select()}} and \code{\link[=relocate]{relocate()}} selections.
}
\examples{
mtcars \%>\% group_by(am, cyl) \%>\% select(group_cols())

}
\seealso{
\code{\link[=groups]{groups()}} and \code{\link[=group_vars]{group_vars()}} for retrieving the grouping variables outside selection contexts.
}
