% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group_vars}
\alias{group_vars}
\title{Return grouping variables}
\usage{
group_vars(x)
}
\arguments{
\item{x}{A \code{data.frame}.}
}
\value{
A character vector of group names.
}
\description{
Return grouping variables
}
\examples{
df <- mtcars \%>\% group_by(am, cyl)
group_vars(df)

}
