% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{n}
\alias{n}
\title{The number of observations in the current group}
\usage{
n()
}
\value{
An \code{integer}.
}
\description{
This function can be used within the context of \code{\link[=summarise]{summarise()}}, \code{\link[=mutate]{mutate()}} and \code{\link[=filter]{filter()}}.
}
\examples{
mt_gears <- mtcars \%>\% group_by(gear)
mt_gears \%>\% mutate(n = n())
mt_gears \%>\% filter(n() < 10)
mt_gears \%>\% summarise(n = n())

}
