% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_distinct.R
\name{n_distinct}
\alias{n_distinct}
\title{Count the number of unique values in a set of vectors}
\usage{
n_distinct(..., na.rm = FALSE)
}
\arguments{
\item{...}{Vectors of values.}

\item{na.rm}{\code{logical(1)}. If \code{TRUE} missing values don't count.}
}
\description{
This is the equivalent of \code{length(unique(x))} for multiple vectors.
}
\examples{
x <- sample(1:10, 1e5, rep = TRUE)
length(unique(x))
n_distinct(x)

}
