% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{peek_vars}
\alias{peek_vars}
\title{Peek at variables in the selection context}
\usage{
peek_vars()
}
\value{
A vector of column names.
}
\description{
Return the vector of column names of the data currently available for selection.
}
