% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{pipe}
\alias{pipe}
\alias{\%>\%}
\title{Pipe}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{The result you are piping.}

\item{rhs}{Where you are piping the result to.}
}
\description{
Pipe
}
\examples{
mtcars \%>\% select(mpg)

}
