% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot-longer.R
\name{pivot_longer}
\alias{pivot_longer}
\title{Pivot data from wide to long}
\usage{
pivot_longer(
  data,
  cols,
  names_to = "name",
  names_prefix = NULL,
  names_sep = NULL,
  names_pattern = NULL,
  values_to = "value",
  values_drop_na = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame to pivot.}

\item{cols}{Columns to pivot into longer format.}

\item{names_to}{The name of the new column that will contain the column
names.}

\item{names_prefix}{A regular expression used to remove matching text from
the start of each variable name.}

\item{names_sep, names_pattern}{If \code{names_to} contains multiple values, this
argument controls how the column name is broken up.
\code{names_pattern} takes a regular expression containing matching groups (\verb{()}⁠).}

\item{values_to}{The name of the new column that will contain the values of
the pivoted variables.}

\item{values_drop_na}{If \code{TRUE}, will drop rows that contain only \code{NA} in the
\code{values_to} column. This effectively converts explicit missing values to
implicit missing values, and should generally be used only when missing values
in data were created by its structure.}

\item{...}{Additional arguments passed on to methods.}

\item{rows_to}{The name of the column that will contain the row names or row
numbers from the original data. If \code{NULL}, will be removed.}
}
\value{
If a tibble was provided as input, \code{pivot_longer()} also returns a
tibble. Otherwise, it returns a data frame.
}
\description{
\code{pivot_longer()} "lengthens" data, increasing the number of rows and decreasing
the number of columns. The inverse transformation is \code{pivot_wider()}
}
\examples{
wide_data <- data.frame(replicate(5, rnorm(10)))
# Customizing the names
pivot_longer(wide_data,
  cols = c(1, 2),
  names_to = "Column",
  values_to = "Numbers"
)

if(require("tidyr")) {
  pivot_longer(
    tidyr::who,
    cols = new_sp_m014:newrel_f65,
    names_to = c("diagnosis", "gender", "age"),
    names_pattern = "new_?(.*)_(.)(.*)",
    values_to = "count"
  )
}


}
