% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poorman-package.R
\docType{package}
\name{poorman}
\alias{poorman}
\alias{poorman-package}
\title{poorman: A Poor Man's Base R Copy of 'dplyr' Verbs}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A simple replication of key 'dplyr' verbs using only base R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nathaneastwood/poorman}
  \item Report bugs at \url{https://github.com/nathaneastwood/poorman/issues}
}

}
\author{
\strong{Maintainer}: Nathan Eastwood \email{nathan.eastwood@icloud.com}

}
\keyword{internal}
