% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull}
\alias{pull}
\title{Pull out a single variable}
\usage{
pull(.data, var = -1)
}
\arguments{
\item{.data}{A \code{data.frame}.}

\item{var}{A variable specified as:
\itemize{
\item a literal variable name
\item a positive integer, giving the position counting from the left
\item a negative integer, giving the position counting from the right
}

The default returns the last column (on the assumption that's the column you've created most recently).}
}
\description{
This is a direct replacement for \verb{[[.data.frame}.
}
\examples{
mtcars \%>\% pull(-1)
mtcars \%>\% pull(1)
mtcars \%>\% pull(cyl)

}
