% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\title{Rename columns}
\usage{
rename(.data, ...)
}
\arguments{
\item{.data}{A \code{data.frame}}

\item{...}{Comma separated key-value paids in the form of \code{new_name = old_name} to rename selected variables.}
}
\value{
A \code{data.frame}
}
\description{
\code{rename()} changes the names of individual variables using \code{new_name = old_name} syntax.
}
\examples{
rename(mtcars, MilesPerGallon = mpg)
rename(mtcars, Cylinders = cyl, Gears = gear)
mtcars \%>\% rename(MilesPerGallon = mpg)

}
