% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rownames.R
\name{rownames}
\alias{rownames}
\alias{rownames_to_column}
\title{Tools for working with row names}
\usage{
rownames_to_column(.data, var = "rowname")
}
\arguments{
\item{.data}{A \code{data.frame}.}

\item{var}{\code{character(1)}. The name of the column to use for row names.}
}
\value{
A \code{data.frame}
}
\description{
Tools for working with row names
}
\examples{
mtcars \%>\% rownames_to_column()

}
