% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice}
\alias{slice}
\title{Choose rows by position}
\usage{
slice(.data, ...)
}
\arguments{
\item{.data}{A \code{data.frame}.}

\item{...}{Integer row values. Provide either positive values to keep, or negative values to drop. The values
provided must be either all positive or negative. Indices beyond the number of rows in the input are silently
ignored.}
}
\description{
Choose rows by their original position in the \code{data.frame}. Grouped \code{data.frame}s use the position within each group.
}
\examples{
slice(mtcars, c(1, 2, 3))
mtcars \%>\% slice(1:3)

}
