% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{transmute}
\alias{transmute}
\title{Transmute}
\usage{
transmute(.data, ...)
}
\arguments{
\item{.data}{A `data.frame`.}

\item{...}{Expressions to mutate the data by.}
}
\description{
Transmute
}
\examples{
transmute(mtcars, mpg2 = mpg * 2)
\dontrun{
mtcars \%>\% transmute(mpg2 = mpg * 2, cyl2 = cyl * 2)
}

}
