% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_all.R
\name{union_all}
\alias{union_all}
\title{Union All}
\usage{
union_all(x, y, ...)
}
\arguments{
\item{x, y}{objects to union all elements of (ignoring order)}

\item{...}{other arguments passed on to methods}
}
\description{
Union all elements of R objects together.
}
\examples{
first <- mtcars[1:20, ]
second <- mtcars[10:32, ]
union_all(first, second)

# union_all does not remove duplicates
a <- data.frame(column = c(1:10, 10))
b <- data.frame(column = c(1:5, 5))
union_all(a, b)

}
