% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{groups}
\alias{groups}
\alias{group_by}
\alias{ungroup}
\title{Group By}
\usage{
group_by(.data, ...)

ungroup(x, ...)
}
\arguments{
\item{.data}{`data.frame`. The data to group.}

\item{...}{One or more unquoted column names to group/ungroup the data by.}

\item{x}{A `data.frame`.}
}
\description{
Determine the groups within a `data.frame` to perform operations on.
}
\examples{
group_by(mtcars, am, cyl)
ungroup(mutate(group_by(mtcars, am, cyl), sumMpg = sum(mpg)))
\dontrun{
mtcars \%>\%
  group_by(am, cyl) \%>\%
  mutate(sumMpg = sum(mpg)) \%>\%
  ungroup()

mtcars \%>\%
  group_by(carb) \%>\%
  filter(any(gear == 5))
}

}
