% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\title{Mutate}
\usage{
mutate(.data, ...)
}
\arguments{
\item{.data}{A `data.frame`.}

\item{...}{Expressions to mutate the data by.}
}
\description{
Mutate
}
\examples{
mutate(mtcars, mpg2 = mpg * 2)
mtcars \%>\% mutate(mpg2 = mpg * 2)
mtcars \%>\% mutate(mpg2 = mpg * 2, cyl2 = cyl * 2)

}
