% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{pipe}
\alias{pipe}
\alias{\%>\%}
\title{Forward-pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{The result you are piping.}

\item{rhs}{Where you are piping the result to.}
}
\description{
Pipe an object forward into a function or call expression.
}
\details{
Unlike the \code{magrittr} pipe, you must supply an actual function instead of just a function name. For example
\code{mtcars \%>\% head} will not work, but \code{mtcars \%>\% head()} will.
}
\examples{
mtcars \%>\% head()
mtcars \%>\% select(mpg)

}
