% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poorman-package.R
\docType{package}
\name{poorman}
\alias{poorman}
\alias{poorman-package}
\title{poorman: A Poor Man's Dependency Free Recreation of 'dplyr'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A replication of key functionality from 'dplyr' and the wider 'tidyverse' using only 'base'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nathaneastwood.github.io/poorman/}
  \item \url{https://github.com/nathaneastwood/poorman}
  \item Report bugs at \url{https://github.com/nathaneastwood/poorman/issues}
}

}
\author{
\strong{Maintainer}: Nathan Eastwood \email{nathan.eastwood@icloud.com}

}
\keyword{internal}
