% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{pull}
\alias{pull}
\title{Pull}
\usage{
pull(.data, var)
}
\arguments{
\item{.data}{A `data.frame`.}

\item{...}{The name of the column to extract}
}
\description{
Pull
}
\examples{
pull(mtcars, mpg)
mtcars \%>\% pull(mpg)
}
