% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relocate.R
\name{relocate}
\alias{relocate}
\title{Change column order}
\usage{
relocate(.data, ..., .before = NULL, .after = NULL)
}
\arguments{
\item{.data}{A \code{data.frame}.}

\item{...}{The columns to move.}

\item{.before, .after}{Destination of the columns selected by \code{...}. Supplying neither will move the columns to the
left-hand side whereas supplying both will result in an error.}
}
\value{
A \code{data.frame}
}
\description{
Change the column positions of a \code{data.frame}.
}
