% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R, R/select.R
\name{rename}
\alias{rename}
\alias{select}
\title{Rename}
\usage{
rename(.data, ...)

select(.data, ...)
}
\arguments{
\item{.data}{A `data.frame`.}

\item{...}{The name(s) of the column(s) to select.}
}
\description{
Rename

Select
}
\examples{
rename(mtcars, MilesPerGallon = mpg)
rename(mtcars, Cylinders = cyl, Gears = gear)
\dontrun{
mtcars \%>\%
  rename(MilesPerGallon = mpg)
}
select(mtcars, mpg, cyl)
\dontrun{
mtcars \%>\% select(mpg)
mtcars \%>\% select(mpg, cyl)
}

}
