% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select}
\usage{
select(.data, ...)
}
\arguments{
\item{.data}{A `data.frame`.}

\item{...}{The name(s) of the column(s) to select.}
}
\description{
Select
}
\examples{
select(mtcars, mpg, cyl)
\dontrun{
mtcars \%>\% select(mpg)
mtcars \%>\% select(mpg, cyl)
}

}
