% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_helpers.R
\name{select_helpers}
\alias{select_helpers}
\alias{starts_with}
\alias{ends_with}
\title{Select Helpers}
\usage{
starts_with(
  match,
  ignore.case = TRUE,
  vars = colnames(get(".data", envir = parent.frame()))
)

ends_with(
  match,
  ignore.case = TRUE,
  vars = colnames(get(".data", envir = parent.frame()))
)
}
\arguments{
\item{match}{\code{character(n)}. If length > 1, the union of the matches is taken.}

\item{ignore.case}{\code{logical(1)}. If \code{TRUE}, the default, ignores case when matching names.}

\item{vars}{\code{character(n)}. A character vector of variable names.}
}
\value{
An integer vector giving the position of the matched variables.
}
\description{
These functions allow you to select variables based on their names.
\itemize{
\item \code{starts_with()}: Starts with a prefix.
}
}
\examples{
mtcars \%>\% select(starts_with("c"))
mtcars \%>\% select(starts_with(c("c", "h")))

}
