% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{slice}
\alias{slice}
\title{Slice}
\usage{
slice(.data, ...)
}
\arguments{
\item{.data}{A `data.frame`.}

\item{...}{A `vector()` of rows to subset.}
}
\description{
Slice
}
\examples{
slice(mtcars, c(1, 2, 3))
\dontrun{
mtcars \%>\% slice(1:3)
}

}
